feat: Update project configuration and remove unused files
This commit updates the project configuration by adding new environment variables to `.env*.local`, modifying the `.gitignore` to include `.nx/workspace-data`, and setting the package manager to `pnpm` in `nx.json`. Additionally, it removes unused files and components from the dashboard app, streamlining the codebase. The Prisma version is also upgraded to ensure compatibility with the latest features.